[XRay] Make llvm.xray.typedevent parameter type match __xray_typedevent
The Clang built-in function is void __xray_typedevent(size_t, const void *, size_t),
but the LLVM intrinsics has smaller integer types. Since we only allow
64-bit ELF/Mach-O targets, we can change llvm.xray.typedevent to
i64/ptr/i64.

This allows encoding more information and avoids i16 legalization for
many non-X86 targets.

fdrLoggingHandleTypedEvent only supports uint16_t event type.

+++ compiler-rt/test/xray/TestCases/Posix/typed-event-logging.cpp
@@ -14,7 +14,7 @@
   printf("after calling the custom logging...\n");
 }
 
-static void myprinter(uint16_t type, const void *ptr, size_t size) {
+static void myprinter(size_t type, const void *ptr, size_t size) {
   assert(type == 42);
   printf("%.*s\n", static_cast<int>(size), static_cast<const char*>(ptr));
 }
